CRS-6: Naive implementation with euclidean distance

+++ src/main/scala/closestRailStation/dataParsing/models/RailStation.scala
@@ -1,3 +1,3 @@
 package closestRailStation.dataParsing.models
 
-case class RailStation(id: String, stationName: String, latitude: Long, longitude: Long) {}
+case class RailStation(id: String, stationName: String, latitude: Double, longitude: Double) {}
